Priniting changed

+++ app/Http/Controllers/front/HomeController.php
@@ -26,8 +26,9 @@ use Carbon\Carbon;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
-use Session;
+use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 
 class HomeController extends Controller
 {
@@ -391,7 +392,11 @@ class HomeController extends Controller
         if ($branch) {
             $sessionId = Session::getId();
 
-            if (Auth::check()) {
+            log::info('session_id', [
+                'session_id' => $sessionId
+            ]);
+
+            if (Auth::check() && Auth::user()->type == 2) {
                 Cart::where('user_id', auth()->id())->delete();
             } else {
                 Cart::where('session_id', $sessionId)->delete();
